KNN.KNN

Documentation

data KNN

Constructors

KNN

Fields

classify :: TestDataCollect -> DataCollect

knn :: Int -> DType -> Resolve -> DataCollect -> TestDataCollect -> DataCollect

overall knn algorithm takes training data,test data, k for knn and distance measure for knn

kNearestNeighbours :: Int -> DType -> DataCollect -> FeatureVector -> DataCollect

finding k nearest neighbours to a test point takes training data,test point, k for knn and distance measure for knn and gives k nearest train points

fitKNN :: Int -> DType -> Resolve -> DataCollect -> KNN

predictKNN :: KNN -> FeatureVector -> DataPoint

predictKNNMulti :: KNN -> TestDataCollect -> DataCollect