NaiveBayes.NaiveBayesInterface

Documentation

crossValidate :: Double -> PriorType -> Distribution -> FPath -> IO Performance

crossValidates the given file and returns the performace of the classfier

traintest :: PriorType -> Distribution -> TrainPath -> TestPath -> String -> IO ()

trains the classifier on the train file and tests it on the test file and writes the classes in a file

makeObjectFromFile :: PriorType -> Distribution -> FilePath -> IO NaiveBayes

Takes the data from a file and fits it with the given priortype and distribution to give IO (naive bayes object)