Structures.Structures

Description

This particular file contains the auxilliary definitions of data that are required by different classifiers

Documentation

type ClassName = String

The name of the class.

type FeatureVector = [Feature]

The feature vector

type Feature = Double

Each of the feature is a double value, the interpretation will be based on the distribution type to be used.

type Probability = Double

just better readability

data DataSet

data set, i.e. the input. each of the feature vector is a point and the classes contain the corresponding class the particular feature belongs to. The size of the points and the classes should be same

Constructors

DataSet

Fields

points :: [FeatureVector]

classes :: [ClassName]

type Probe = ClassName -> Feature -> Probability

Probe is basically a function that gives the probability of a feature having a particular value for a particular className

type FeatureData = [Feature]

this is same feature for different input points

type ClassData = [ClassName]

this is the parallel data for classes corresponding to the input

type TestDataCollect = [FeatureVector]

Collection of feature vectors

data DistributionType

specifies the type of the distribution used to fit the training data for a particular feature Source of Description : MathWork Normal - Gaussian Mvmn - Multivariate multinomial distribution Mn - Multinomial distribution

Constructors

Normal

Kernel

Mvmn

Mn

data Distribution

Constructors

Single DistributionType

Multiple [DistributionType]

data PriorType

This defines which prior probablity to use. Empirical - Depends on the training data. Uniform - Equal probablity for all classes Specific - User provided probability

Constructors

Empirical

Uniform

Specific [(ClassName, Probability)]

type DataPoint = (FeatureVector, ClassName)

data point, is the combination of point and its class data DataPoint = DataPoint { point::FeatureVector ,className :: ClassName }

type DataCollect = [DataPoint]

alternative to DataSet but in this classes or points are not directly separable

data DType

distance measure type DType = String

Constructors

Euclidean

Manhattan

DefaultDT

data Resolve

Resolution method

Constructors

Last

First

DefaultR

type NeighbourNum = Int

The number of neighbours the KNN uses

type ConflictControl = String

ROMIL

type Label = String

NITESH

data C45Dtree

Constructors

C45Dtree

Fields

c45Classify :: Datum -> Label

data DecisionTree

Constructors

Leaf (Maybe Label)

Node

Fields

attribute :: Attribute

decide :: Feature -> DecisionTree

newtype Datum

Constructors

Datum

Fields

attributes :: [(Attribute, Feature)]

data Labelled

Constructors

Labelled

Fields

label :: String

point :: Datum

type DataSetDT = [Labelled]

type FeatureDT = Datum

data Attribute

Constructors

Nominal

Fields

attName :: String

possibleValues :: [Feature]

Numeric

Fields

attName :: String

data DataSetGen

General definition of Dataset to unify the interface

Constructors

DT_NB DataSet

DT_KNN DataCollect

DT_DT DataSetDT